Refactor the packaging
